+++ navajo-cli/src/lib.rs
@@ -2,8 +2,10 @@ mod aad;
 pub mod algorithm;
 mod cli;
 mod encoding;
+mod segment;
 
 pub mod envelope;
 pub use aad::*;
 pub use cli::*;
 pub use encoding::*;
+pub use segment::Segment;