+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+-   Added
+-   Removed CLI alias `"e"` for enable-key and `"d"` for disable-key
 -   Renames DSA keyring `"key_pair"` to `"value"` (#52)
 -   Removes async from `navajo-cli`, making it sync for easier testing & resolving (#46)
 -   Each primitive's `KeyInfo` were renmaed from {Primitive}KeyInfo to `KeyInfo` (#53)